pyproject.toml
[tool.poetry]
name = "homework29"
version = "0.1.0"
description = ""
authors = ["puzynjailya <puzynjailya@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
psycopg2 = "^2.9.3"
djangorestframework = "^3.13.1"
Pillow = "^9.2.0"
djangorestframework-simplejwt = "^5.2.0"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.5.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
